W3TC's new nginx.conf on fresh install -- Fixes #9
The issue was not with my new `include` reference to W3TC’s editable
nginx.conf, at WordPress root, in the main VHost, but with W3TC
settings themselves as they were in its old nginx.conf file.  This
fixes #9 but we still need to configure W3TC to work optimally.